+++ packages/framework/foundation-compose/src/compose.tsx
@@ -8,7 +8,7 @@ import {
   IExtractState
 } from '@uifabricshared/foundation-composable';
 import { initializeStyling, getOptionsFromObj } from './useStyling';
-import { immutableMerge } from '@uifabricshared/immutable-merge';
+import { immutableMerge } from '@fluentui-react-native/immutable-merge';
 import { ISlotProps } from '@uifabricshared/foundation-settings';
 
 /**